requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
PyQt6>=6.6.1
PyQt6-Qt6>=6.6.1
PyQt6-sip>=13.6.0
Pillow>=10.2.0
obsws-python>=1.0.0

# Development dependencies
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0 